Document the 'RestClient::RequestFailed' error
Whilst it is documented in the Developer Docs (https://docs.publishing.service.gov.uk/manual/alerts/high-priority-tests.html), this is a common enough error that I believe it's worth a place in the README. The error happens whenever you cancel a running test suite, which you're likely to do if you notice the first test has failed.

This commit also adds a shortcut alias to take some of the pain out of process management.