Fix 'killbrowsermob' alias
Command worked on the basis that the browsermob process would be the first thing in the grepped output, and the grep command itself would be the second thing. But when stored as an alias, the grep command is the first thing and the browsermob process is the second thing.

Changing browsermob -> [b]rowsermob ensures that the browsermob process is still found ("find the first b followed by rowsermob"), but that the grep process itself isn't matched.

Secondly, the \$2 was getting evaluated too early, so the ID wasn't being grabbed properly. That is now escaped."