Re-order search analytics tests to check at the end
These tests are flaky, but I've not been able to replicate that
locally.  Failures seem to often be when checking for the
"filterClicked" event, so I suspect there's not enough time between
the search options being expanded and the check for the request, as
the request is asynchronous.

So I've just moved all the checks to the end.