Simplify tests_json_output.sh
This script doesn't manage the cucumber process properly, so just
remove that as a concern, Upstart can do this. Also, use exec to avoid
having the bash process stick around, and use the --out option for
cucumber to write the output file, as this avoids redirecting the
output through bash.