Remove absolute check on DGU datasets
Previously we experienced an issue where a significant number of
datasets disappeared from the DGU Find frontend due to a suspected API
discrepancy from CKAN. We added two features to detect this issue in
future, but one requires manual intervention to keep the check accurate.

This removes the accuracy feature, since we believe it is covered by the
other feature we added to check the agreement between the number of
datasets reported by CKAN vs. DGU Find. In particular, there is no
evidence that the total count of datasets reported by CKAN was at fault.

https://github.com/alphagov/smokey/pull/493