adds test to ensure default web acl rule in place
the default web acl in place on the cache ALB has a rule that blocks
traffic if the header X-Always-Block: true is set on a request. We can
use this as an basic sanity check that the waf is associated with the
ALB and as a initial example test for others to follow as more rules are
added.