Don't show an unknown if there are no tests
This partially reverts e12e983b46d483d9d595bdd704f5c7964fc678b0 to
restore the previous behaviour of sending an OK status to Icinga when no
tests were run (rather than an UNKNOWN which I had changed it to). This
is because when no tests are run it means the feature has no tests at
the priority, not that there is a problem.