Suppress JS error for YouTube embed
There is a long standing bug [1] in the YouTube embed via their API.

The page will still work for users, as will the video, but the
Javascript error is causing Smokey to break. We have tried various
suggested mitigations and fixes but none have worked.

As the test is only really testing the draft stack and not videos,
switching out the test page for one without a Youtube video seems a
sensible change.

[1] https://issuetracker.google.com/issues/35174286