Update tests for retiring stats finder
The Whitehall statistics finder is soon to be redirected to the newer
https://gov.uk/search/research-and-statistics finder. I've updated tests
that will fail when that redirect goes in.

There is one remaining test that should be updated (the Whitehall tests for
announcements and statistics atom feeds).  This test will continue to pass
once the redirects are deployed, but should be tweaked to ensure that the
redirects are actually working.