Fix for failing Smokey test
I believe as part of a general modernisation of the use of govuk_frontend
we're seeing a test fail in Smokey due to being unable to find a certain
css field. This PR changes the means in which we retrieve the object in
question from the now non-existent css id to the id as found in the page
source.