Run finder benchmarking tests in AWS only
We occasionally get a failure for these tests when they are run in carrenza,
which we think is related to network latency.  We've checked the finder-frontend
application logs and the maximum time to respond hasn't exceeded 1.6 secs for
the research and statistics finder in this period.

These tests should be failing if the software is slow, not if there are
problems with network connectivity.