Remove email signup steps for double opt-in
This removes a couple of steps from one of the email signup scenarios,
which went beyond the 'signup' journey and through the 'subscribe'
journey, which will shortly change with the introduction of double
op-in. The interaction between double opt-in and a received email means
we can't preserve an E2E smokey test for subscriptions, but we're
comfortable this is covered by the of specs in the respective repos.