Set Chrome user agent to Smokey Test / Ruby
Our PAs find it tricky to omit Smokey traffic from our analytics.  This is
more important than it used to be because of our reduced analytics traffic.

This sets the Chromedriver user agent to be the same as the user agent used
by RestClient for some of the other Smokey requests.