Fix ambiguous search result error
As we show parts in search results now, this step can fail if a part
has the same title as a normal result, as capybara doesn't know which
to click.

So the solution is to call `click` on the right element directly,
keeping it to proper results (not parts) by adding the class: parts
are a .gem-c-document-list-child__link, and so don't match.