Run 'cucumber' directly from jenkins.sh
Previously we defined a rake task for each environment in which we
run Smokey, which is a confusing layer of indirection. Currently,
the Jenkins job that runs smokey sets two environment variables:

- TARGET_PLATFORM e.g. integration, staging_aws
- MYTASK e.g. test:integration, test:staging_aws

This replaces the rake tasks by using the 'TARGET_PLATFORM' variable
to set the 'ENVIRONMENT' and cucumber profile, which is all the tasks
were doing previously. Removing the layer of indirection will also
make it easier to pass additional flags to the Cucumber CLI.