Allow optional scoping by application
This modifies jenkins.sh so it can further restrict the scenarios
to be run when an optional 'APPLICATION' environment variable is
set. Initially, we only need this functionality for smart-answers,
as part of moving to Continuous Deployment. The 'app-' prefix is
used to avoid ambiguity between with a 'theme' e.g. 'app-frontend'
is clearer than just 'frontend'.