Set basic authentication at proxy level
This sets the basic authentication at the point of making any HTTP
request rather than just at the opening URL. This is being put in to
resolve a problem where we integration is failing on a test to view a
CSV preview because it lacks the basic auth needed to access the assets.
This is because we're in an odd situation where the page is hosted on
the assets hostname and the assets are hosted on the www one.