Treat all scenarios with the same priority
Previously the script would match test output against the priority
specified by Puppet, for each priority it alerts for [1]. This changes
the script to only report all test results when called with "normal",
ignoring any priority tags associated with each scenario. When the
script is called with any other priority, it will now return nothing.

This makes the "priority" redundant.

[1]: https://github.com/alphagov/govuk-puppet/blob/4ec2949332079570be5493753bb84a983218a444/modules/icinga/manifests/check_feature.pp