Tag scenarios for Continuous Deployment apps
This adds "@app-<app>" tags for features and specific scenarios that
test whether one of the following apps is functioning:

- Collections
- Travel Advice Publisher
- Publisher
- Finder Frontend
- Government Frontend

In future work, we may want to combine scenarios for the same app into
a single feature file, instead of separating them by feature.

Note that publisher apps are primarily tested via the Publishing E2E
tests; the number of scenarios in this repo is limited.