Add /healthcheck scenarios for 5 apps
This adds basic /healthcheck scenarios in preparation for enabling
Continuous Deployment for 5 more apps, following the approach we
took for Smart Answers and Publishing API [1][2].

In future work [3], we expect to replace the /healthcheck scenarios
in favour of the Continuous Deployment pipeline checking the existing
Icinga alerts for this endpoint.

Note: Publisher does not have a /healthcheck endpoint.

[1]: https://github.com/alphagov/smokey/pull/696
[2]: https://github.com/alphagov/smokey/pull/681
[3]: https://trello.com/c/V1xnRofx/178-use-icinga-servicegroups-to-replace-smokey-healthcheck-scenarios