Specify finder-frontend tests to run against AWS
This commit specifies via the `@aws` tag to run the `@app-finder-frontend` tests against AWS, as that is where the app is served from. The reason for this is to fix the Smokey test for the finder-frontend healthcheck which is currently resolving against what would have been the public endpoint on Carrenza and which now no longer exists. This test has been consistently failing since it was introduced, and was not marked to test against AWS when other apps hosted on AWS were (see https://github.com/alphagov/smokey/commit/91a5ce94b7b96d09f5db87e63fbc3e400788de00).