Suppress JS console errors in Licensify tests
We're currently unable to deploy changes to Licensify. Our
Smokey test suite is failing because we're seeing CSP errors
on Google Analytics requests in Licensify.

In order to get Smokey passing again - essential for continuous
deployment and for developer confidence - we've decided to
ignore the JS error for these tests. When we're able to
deploy Licensify again, we should already have fixed the CSP
issue here:
https://github.com/alphagov/licensify/pull/492