Re-enable malicious code test for finder-frontend
Previously this test was disabled [1] over a month ago due to being
flakey. We believed that there was an issue with CORs and google
analytics which led to a card being created to investigate. Since then
I've tried re-enabling the test to replicate the flakeyness for
integration and staging which proved to be unsuccessful, as out of 30
runs, there were no failures. This commit re-enables the test as there
doesn't seem to be an issue anymore.

[1]: https://github.com/alphagov/smokey/pull/755