Remove carrenza-specific notaws flags from cucumber profiles
This is good, since now we can use staging rather than staging_aws
as the profile. It'll be less confusing, since we no longer use
Carrenza.

The notaws flag is not used anymore except in this file.

I've left in staging_aws and production_aws, as these are still
in use. Once we've stopped using these we can remove these profiles.