Mark Healthcheck test as failing for replatforming
Smokey currently doesn't have access to frontend
in ECS, so this will fail until we fix that.

Given Plek is used to determine the internal URL
for an app we could set PLEK_FRONTEND_... on our
smokey container in ECS. However, I'm not sure we
need the smoke tests to check the health check
endpoint of a new container given ECS will do this
for us.