Change path of X-Always-Block test
We're now ignoring paths in the cache key for the
homepage. The WAF is behind Fastly so the requests
from this smoke test to the homepage won't actually
make it to the WAF.

The fix is to point the smoke test at another path.

Test it locally:

```
ENVIRONMENT=production_aws bundle exec cucumber \
--profile production_aws --strict-undefined features/waf.feature
```