Only run Publisher publishing tests in integration
Smoke tests were recently added to test the end-to-end Publisher workflow - specifically creating, publishing and unpublishing content. In order to facilitate the publishing test, Smokey was granted the `skip_review` permission in integration and staging which then enables it to request 2i, skip review and publish immediately. Due to the daily Signon data sync from prod to staging however, manually added permissions do not persist in staging; as Smokey doesn't hold the `skip_review` permission in prod, this will also be the case in staging.

For now, the Replatforming team are deciding to only run this test in integration, where manually added permissions in Signon do persist as there is no daily Signon data sync that would otherwise overwrite this data. When we come to moving the new platform to staging, we will revisit this decision and consider doing the work to persist Smokey's `skip_review` permission for Publisher in staging.