Add tests for signing in to the GOV.UK account
We've had a few instances where the sign in journey has broken.  This
sort of thing is hard to catch when testing at the app level, as it
involves GOV.UK apps, an external auth service, and our CDN config.

We think that smoke tests which test the full journey from the
perspective of a user will help catch future breaking changes.