Fix healthcheck for authenticating-proxy
This was making a request to the wrong app. Note that draft-origin
doesn't have a 'govuk-internal' DNS name, so we need to use the
public one. In future work we should consider giving the proxy its
own domain name so we can treat it consistently like other apps.
(We didn't need to previously because Varnish makes request directly
to its port on the same machine [1].)

[1]: https://github.com/alphagov/govuk-puppet/blob/32c1bbbb10067078c1406170666a135b4a10aaea/modules/govuk/manifests/node/s_cache.pp#L95