Add Smoke Tests for Get Involved Page
As part of the migration from Whitehall to Government Frontend it was
realised that we had little in the way of testing the page. This smoke
test is designed to check all static and dynamic elements in the page,
deemed necessary due to the amount of cross-stack reliance the page
now has.