Move 'Get Involved' tests to government-frontend app tests
When run through the continuous deployment pipeline, only the tests for
the specified app are run. When government-frontend has been deployed
recently, the 'Get Involved' tests never ran. As the change in [1]
caused these tests to fail but the tests never ran in CD, the change was
deployed to production and we only noticed when the smokey loop failed
in production.

Moving these tests to the government-frontend app will ensure they run
when after government-frontend is deployed to each environment.

1: https://github.com/alphagov/government-frontend/pull/2284