Fix step in Get Involved test
An unused div was removed from the Get Involved page in [1] in order to
fix the `And it should be populated` step of this test.

However this now causes the `When I click the next closing response
link` step to fail, as the div is no longer present.

Updating the test to match the government-frontend changes.

1: https://github.com/alphagov/government-frontend/pull/2296