Upgrade browserup-proxy
This follows the pattern of dc6b166c3ade8b16793add4818088d1c2add8ca0 -
copying all the jars and the bin/browserup-proxy file into the smokey
repo.

This version is built from
https://github.com/richardTowers/browserup-proxy/tree/upgrade-log-4j-2-1-2
which is a fork of version 2.1.2 which also updates the version of log4j
to 2.16.0 (which fixes a critical vulnerability).

I built browserup-proxy using:

./gradlew clean build distZip

This creates a zip in browserup-proxy-dist/build/distributions/browserup-proxy-2.1.2.zip

I then unzipped that, and copied the files from lib into lib and bin
into bin.