Remove `webdrivers` gem dependency
The `webdrivers` gem is no longer required because supported GOV.UK runtime environments now supply a ChromeDriver for use by Selenium – e.g. govuk-docker and Jenkins CI workers.

Smokey also runs as a containerised application in the new EKS hosting environment that's currently being built. For that, I've updated the Dockerfile in this repository to install Chromium and a compatible ChromeDriver from the default package repo. That way, we'll be guaranteed to have a compatible browser and webdriver installed, and it should be possible for developers to run it on an M1 MacBook because Chromium supports both AMD64 and ARM64.