Remove unnecessary @benchmarking tag
This stops the scenarios in benchmarking.feature from running when
we do a full run of Smokey. However, the tests are still run by the
Smokey Loop. Since we're removing the single alert associated with
a full run of Smokey [^1], the impact of these tests failing is low
and we may as well just run them.

Long term we should reconsider the need for these specific tests:
it's unclear why most of them exist.

[^1]: https://github.com/alphagov/govuk-puppet/pull/11610