Trial alerting on scenario failures in Sentry
Problem: we think Smokey is flakey but lack the data on what / why.
Flakiness was a common issue identified in the last dev tools survey.
We need to understand the flakiness if we want to reduce it.

Possible solutions:

- Metrics in Grafana*. This would help us understand what is failing
but not why, which isn't very useful overall.

- Logs in Kibana**. This would help us understand why a scenario is
failing but lacks good high level analysis tools.

- Sentry (this PR).

If we think of the Smokey Loop as another app - it does run like an
app - then Sentry is a potential option that combines the "what" and
the "why" data to help us understand what is flakey.

How to structure the events
===========================

Unlike a normal app, we probably don't want distinct events for every
scenario or type of error, as this makes the main incident page long
or harder to understand at a glance. Instead, I suggest we group each
error by feature by forcing this as the "fingerprint".

Despite the forced grouping, we can add some "tags" to each event to
support more fine-grained searches. This is useful if:

- Multiple scenarios have the same step and are failing for the same
reason.

- Multiple scenarios in the same feature are failing and we want to
isolate each one.

What is the next step for this?
===============================

- Decide if this is a direction we want to go (pros / cons).
- Double check if Sentry rate limiting is a concern.
- Setup a real project for Smokey in Sentry.
- Update Puppet to populate the SENTRY_DSN env var for Smokey Loop.
- Let events accumulate for a week to see (at last) what is flakey.

Notes about Grafana / Kibana
============================

*We do have metrics from Icinga about Smokey Loop failures [^1], but
the metric only updates when Icinga's state changes e.g. when alerts
go from "OK" to "CRITICAL". This is useful if a scenario is flapping
a lot, but not if it has a sustained failure.

**We do have logs in Kibana about Smokey Loop failures [^2], but the
logs come from syslog and aren't readable. It would be hard to make
a log that was readable but also indexed in such a way as to give a
high level picture - for monitoring.

[^1]: https://grafana.integration.publishing.service.gov.uk/dashboard/file/icinga_alerts.json?orgId=1
[^2]: https://kibana.logit.io/s/d4dba58d-d7be-4bca-81e7-034e14b8b9ee/app/kibana#/discover?_g=()&_a=(columns:!(message),index:'*-*',interval:auto,query:(language:lucene,query:'failed%20AND%20skipped'),sort:!('@timestamp',desc))