Try moving blacklist URLs to ignored errors
The blacklisting meant we didn't actually know if requests to GA
were successful [^1]. In terms of whether that use case is real:
bots crawl GOV.UK all the time, so we shouldn't be concerned about
a small amount of Smokey traffic.

The other use cases are only concerned with ignoring errors, which
we can equally do with log filters.

[^1]: https://github.com/alphagov/smokey/blob/71f157fa17efe9d561757051975844d794719659/features/step_definitions/analytics.rb#L29