Remove disabled Account API feature
This only ever ran on Staging [^1][^2] and is unnecessarily tying us
to the BrowserUp Proxy. I've checked with #govuk-accounts-tech and
with #find-and-view-tech and there are no plans to reinstate it.

If we do reinstate this feature, we should invest effort in finding
another way to manage access without needing basic to use a proxy -
after all, these tests are meant to emulate real user behaviour.

[^1]: https://github.com/alphagov/smokey/pull/868
[^2]: https://github.com/alphagov/smokey/pull/866