Use Chrome performance logs to check AJAX requests
While these assertions are dubious compared to the user-focussed
style of the rest of the tests, we can at least avoid the proxy.
Local testing suggests the new approach is a few seconds faster,
probably because we're not waiting for Java to start.

I've also taken the opportunity to make the analytics assertions
more robust by checking the URL.