Add Sentry / StatsD reporting for scenario errors
Tested and works with errors in individual steps, as well as with
the "After" hook to detect JS errors. Backtraces for both types of
error usually include the step and its Ruby definition; we also tag
the feature and the scenario to help identification and filtering.
Sentry does a good job of expanding backtraces to show all the code.

Using govuk_app_config adds significant value:

- Ensures Sentry config is consistent with the rest of GOV.UK.
- Includes issue filtering by environment and for data sync errors.
- Pushes extra stats to StatsD for monitoring by exception type.

I think this compensates for the handful of redundant gems, such as
railties, that get pulled in with this gem.

As mentioned in the comment, we need to use "capture_message" instead
of "capture_error". This makes little difference in Sentry as far as
I can see, except that we need to manually populate the hint [^1][^2].

[^1]: https://github.com/getsentry/sentry-ruby/blob/7848d59450ae6bcb8dea6c141abdcdfe7b218567/sentry-ruby/lib/sentry/hub.rb#L100
[^2]: https://github.com/alphagov/govuk_app_config/blob/87e445eccee5fba2449a170d5ba628e8a380fcb8/lib/govuk_app_config/govuk_error/configuration.rb#L109