Add guidance for installing Ruby / Chromedriver
This repo isn't covered by GOV.UK Docker, which takes care of both
of these setup concerns. My quick attempt at adding it was partly
successful, but when running the tests the latter ~half fail e.g.

    Then I should be able to visit:                 # features/step_definitions/smokey_steps.rb:162
      | Path                           |
      | /government/announcements.atom |
      | /government/statistics.atom  |
      invalid session id (Selenium::WebDriver::Error::InvalidSessionIdError)

For now it makes sense to at least offer guidance on running this
repo natively. Hopefully we can add it to GOV.UK Docker in future.