Fix logs from scenarios leaking into others
Steps to reproduce:

- Create a fake scenario that loads a page with JS errors but fails
normally. Something like this:

    When "whatever" do
      visit "page-with-js-errors"
      raise "fail and avoid checking JS errors"
    end

- Run that scenario followed by another scenario that normally passes
and see that it fails with the JS errors.

Sometimes JS errors may be unrelated to the scenario, so we check them
even if the scenario passes [^1]. This has the effect of clearing the
browser logs for the next scenario. However, if a scenario fails, we
don't do this check / flush and the logs leak into the next scenario.

This commit fixes the bug by explicitly flushing and clearing browser
logs before each scenario runs, ensuring we have a clean slate.

[^1]: https://github.com/alphagov/smokey/blob/00af2dc33a6890c1960092c488bdaa16ec267a79/features/support/hooks.rb#L13