Fix commented out AB test scenario
This seems to have been implemented separately to the general ways
we intercept AJAX requests (previously a proxy and now via Chrome).

I had to do a little analysis to figure out the actual string that
is present in the request params, which looks like this:

    https://www.google-analytics.com/j/collect?v=1&_v=j96&aip=1&a=1204582135&t=pageview&_s=1&dl=https%3A%2F%2Fwww.integration.publishing.service.gov.uk%2Fhelp%2Fab-testing%3Fsmokey_cachebust%3D0.2070604651042014&dp=%2Fhelp%2Fab-testing%3Fsmokey_cachebust%3D0.2070604651042014&ul=en-us&de=UTF-8&dt=A%2FB%20testing%20-%20GOV.UK&sd=24-bit&sr=800x600&vp=785x600&je=0&_u=SCEACAABBAAAAC~&cid=393051546.1654608380&tid=UA-26179049-22&_gid=324170629.1654608380&_slc=1&cd15=200&cd16=unknown&cd95=393051546.1654608380&cd11=1&cd2=special_route&cd3=other&cd4=a4d4e755-3d75-4b19-b120-a638a6d79ba8&cd12=not%20withdrawn&cd17=special_route&cd20=frontend&cd30=none&cd31=none&cd32=none&cd56=other&cd57=other&cd58=other&cd59=other&cd39=false&cd89=frontend&cd26=0&cd27=0&cd23=unknown&cd38=false&cd40=Example%3AB&cd100=true&z=1231930786

Although we use CGI::escape elsewhere that doesn't work here as it
encoded "=" and ":". Only the ":" is encoded in practice so for the
sake of simplicity I've just done this manually - I don't think it's
worth making such a small number of steps consistent with encodings.