Make testing the "full stack" the default
This just means we're visiting pages via the "gov.uk" domain and
not via internal or app-specific domains. It's a sensible default
as the majority of tests had it set and forgetting to set it could
lead to subtle bugs if previous test state is still present.

See features/support/hooks.rb and features/step_definitions for the
main changes in this commit.