Speed up sitemap test by only checking one link
The sitemap is a critical feature of GOV.UK. However, we really,
_really_ don't need to check the whole thing to prove it works.

Before this change, the scenario would take ~40s to run, which is
about 25% of the total runtime of all Smokey tests - this was by
far the slowest test in the whole suite. Now it will only take ~3s
to run - a significant improvement.