Consolidate CDN layer tests into cdn.feature
While the caching tests also pertain to Varnish at the origin, we
may not be using Varnish forever so the CDN seems a better fit.

I've tweaked some of the scenario names so they still make sense
in their new home. I've also pushed tags onto scenarios where they
don't apply to all the scenarios that are now in the same file.