Consolidate origin layer tests into origin.feature
The redirect test could arguably go in router.feature, but it's a
site-wide behaviour so it also fits well in here.

I've tweaked some of the scenario names so they still make sense
in their new home. I've also pushed tags onto scenarios where they
don't apply to all the scenarios that are now in the same file.