Break up "Check the feedback component loads"
This is targeting individual apps, not the feedback app. If this
feature is truly critical for all of the apps, then we should be
willing to have it duplicated in each of them*.

All of the other frontend features have a "@replatforming" tag,
so I've repeated this pattern for email-alert-frontend.

*I don't actually believe this is true and would be open to us
deleting some of these scenarios e.g. I don't think it's truly
critical for users to give feedback about info-frontend.