Improve test for visiting a draft asset
Previously we would try to view a non-existent asset, which didn't
matter as we were only checking the redirect to Signon, which works
irrespective of the asset existing [^1].

Now we test the full journey of viewing a draft asset*. I've chosen
to use an image because this will preview in the headless browser,
which is easier to check and avoids an actual download.

*Note: the asset is published but, like published documents, also
exists in draft as well.

[^1]: https://github.com/alphagov/asset-manager/blob/1e88cd5b71b8ac6f73a079176b9f74c57b608d1c/app/controllers/media_controller.rb#L54