Remove signon secure cookie test
This test is being removed because it doesn't meet the eligibility
criteria for this repository.

I investigated whether it could be imported to signon, but this is not a
simple task as the "secure" property is defined by the environment the
code is executing in [1].

With this knowledge I decided we could delete this for a few reasons:

1) It's unconventional for us to test convention in this way
2) The risks of this cookie not being secure are very low due to HSTS
   making it very difficult to visit the site without HTTPS
3) There are various cookie tests in signon, which make the likelihood
   of a regression low

If we were to want to test this Smokey could be the place if we were to
define an eligibility criteria of environment configuration. As this
test seems to be the lone case of this need, this step seems
superfluous.

[1]: https://github.com/alphagov/signon/blob/e2d48a22a2afe1842cf88921da67490f04582cf2/config/initializers/session_store.rb#L5
[2]: https://github.com/alphagov/signon/blob/e335543918e9e1e189a8791f26a85de3a39dd534/test/integration/cookies_security_test.rb#L4-L11