Enable network by default
I can't see the benefit of running Smokey unless we're making
actual network requests.
In practice, we always DO use network requests, because we have
a `RATE_LIMIT_TOKEN` ENV var defined for each Smokey run. So it
isn't clear when we'd ever run Smokey without this variable set.

If we do remove the variable, network requests are stubbed (I
believe), meaning we'll get a false positive from smokey.feature,
where we have a test that Smokey can successfully make a network
request.

The conditional is still worth keeping as we should handle the
case where the rate limit token is missing, but I think it's a
fair assumption that only real network requests should be used.