Change feedback close button text to cancel
In a recent change to sitewide feedback form, the text for the close
button was changed from 'Close' to 'Cancel'. This resulted in "I close the open
feedback form" to fail because it was clicking on a button with "Close"
text. The label of this button has been updated in the test, so the
button can now be found.