Replicate the publishing-e2e-tests setup for Chrome
The e2e tests are now retired, and were not looked upon very
favourably for quite some time, but it has to be said that their
Dockerfile ([1]) and Selenium webdriver setup ([2]) appeared to be
of a higher standard than what existed in Smokey.

This commit is an attempt to modernise our headless browser setup, in
order to fix CI, which has started consistently failing.

We achieve this by:

1) Updating the Dockerfile
2) Tweaking the env.rb file

It's interesting to note that it is the GitHub Actions CI jobs that
have started failing regardless of what code changes are made, e.g.
https://github.com/alphagov/smokey/pull/1046

And even more interesting to note that the bump to selenium-devtools
fixes the issue for GitHub Actions:
https://github.com/alphagov/smokey/pull/1041
However, it introduces its own issues on Jenkins:
`unknown error: Chrome failed to start: crashed`

Therefore the manual changes in this commit are required.

Note that I also had to move the rate limit header logic into the
'http_requests.rb' helper file. Trying to set this at the web driver
level is a fools errand:

https://stackoverflow.com/questions/15645093/setting-request-headers-in-selenium

[1]: https://github.com/alphagov/publishing-e2e-tests/blob/5c2a09b49e2bd4dae1d2feaaad82bd65a4efa722/Dockerfile
[2]: https://github.com/alphagov/publishing-e2e-tests/blob/5c2a09b49e2bd4dae1d2feaaad82bd65a4efa722/spec/spec_helper.rb#L119-L132